backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
Faker==30.6.0
python-dateutil==2.9.0.post0
motor==3.6.0
pymongo==4.9.1
pydantic==2.9.2
python-dotenv==1.0.0
kafka-python==2.0.2
langgraph==0.2.45
langchain-core==0.3.27
langchain-openai==0.2.8
langchain-google-genai==2.0.9
